Updated StopWatchButton Functionality and Userface design
