+++ public/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
     <head>
-        <title>Template Site</title>
+        <title>Tony's StopWatch</title>
     </head>
     <body>
         <section id='root'></section>
