+++ src/StopWatchButton.tsx
@@ -1,12 +1,17 @@
 import React from 'react'
 
+
 interface Props{ 
     name:string,
-    onClick: () => void;
+    onClick: () => void,
+    className: string
+  
 }
 
-export default function StopWatchButton({name, onClick}:Props) {
+export default function StopWatchButton({name, onClick, className}:Props) {
     return(
-        <button onClick={onClick}> {name} </button>
+        // Leveraging Props, we enhance the ability to customize individual buttons 
+        // and manage onclick event handlers more effectively.
+        <button onClick={onClick} className={className}> {name} </button>
     )
 }
